tests/wpt/tests/credential-management/fedcm-csp.https.html

<!DOCTYPE html>
<title>Federated Credential Management API network request tests.</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<body>

<script type="module">
import {request_options_with_mediation_required,
        fedcm_test,
        set_fedcm_cookie} from './support/fedcm-helper.sub.js';

fedcm_test(async t => {
  const cred = navigator.credentials.get(request_options_with_mediation_required());
  return promise_rejects_dom(t, "NetworkError", cred);
}, "Provider configURL should honor Content-Security-Policy.");

</script>
